+++ src/main/java/com/example/projects/dto/PointDTO.java
@@ -2,24 +2,24 @@ package com.example.projects.dto;
 
 public class PointDTO {
 
-    private Long id;
+    private int id;
     private double x;
     private double y;
 
     public PointDTO() {
     }
 
-    public PointDTO(Long id, double x, double y) {
+    public PointDTO(int id, double x, double y) {
         this.id = id;
         this.x = x;
         this.y = y;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
